requirements.txt
fastapi==0.98.0
uvicorn==0.22.0
python-multipart==0.0.6
opencv-contrib-python==4.7.0.72
numpy==1.25.0
python-dotenv==1.0.0
dlib==19.24.2
face-recognition==1.3.0
uuid0==0.2.7